strt

+++ program/src/include/funct.h
@@ -4,12 +4,12 @@
 
 void form_time(struct AppParams ap_pr, float* t);
 void form_Uvx(struct AppParams ap_pr, float* Uvx, float* t);
-void form_Uvix(struct AppParams ap_pr, float* Uvx, float* Uvix);
-void form_tabl1(int n, float *t, float *Uvx, float *Uvix);
+void form_Uvix(struct AppParams ap_pr, float* Uvx, long double* Uvix);
+void form_tabl1(int n, float *t, float *Uvx, long double* Uvix);
 
-float parametr(int n, float sum, float *U, float *t);
+float parametr(int n, float sum, long double *U, float *t);
 
-void file_out_data(int n, float* t, float* Uvx, float* Uvix);
+void file_out_data(int n, float* t, float* Uvx, long double* Uvix);
 
 void control_calc(struct AppParams ap_pr);
 void approx_value(struct AppParams ap_pr);
